refactor: componentsディレクトリをlayout/sections/uiに整理
