+++ app/components/sections/KeyVisual/FeatureBadge.tsx
